+++ packages/utils/src/commons/get-is-server.ts
@@ -1,4 +1,4 @@
-import { getIsClient } from '@/common/get-is-client.js';
+import { getIsClient } from '@/commons/get-is-client.js';
 
 /**
  * 호출된 환경이 서버 환경인지 여부를 반환
