+++ packages/utils/src/commons/is-empty.ts
